requirements.txt
pandas>=2.0
pdfplumber>=0.11.0
tabula-py>=2.9.0
jpype1>=1.5.0
